+++ src/k8s-extension/azext_k8s_extension/partner_extensions/AzureMLKubernetes.py
@@ -30,7 +30,7 @@ from .PartnerExtensionModel import PartnerExtensionModel
 
 logger = get_logger(__name__)
 
-resource_tag = {'created_by': 'amlk8s-extension'}
+resource_tag = {'created_by': 'Azure Arc-enabled ML'}
 
 
 class AzureMLKubernetes(PartnerExtensionModel):
@@ -271,10 +271,10 @@ def _lock_resource(cmd, lock_scope, lock_level='CanNotDelete'):
     lock_client: azure.mgmt.resource.locks.ManagementLockClient = get_mgmt_service_client(
         cmd.cli_ctx, azure.mgmt.resource.locks.ManagementLockClient)
     # put lock on relay resource
-    lock_object = ManagementLockObject(level=lock_level, notes='locked by amlk8s.')
+    lock_object = ManagementLockObject(level=lock_level, notes='locked by amlarc.')
     try:
         lock_client.management_locks.create_or_update_by_scope(
-            scope=lock_scope, lock_name='amlk8s-resource-lock', parameters=lock_object)
+            scope=lock_scope, lock_name='amlarc-resource-lock', parameters=lock_object)
     except:
         # try to lock the resource if user has the owner privilege
         pass
@@ -337,7 +337,7 @@ def _get_service_bus_connection_string(cmd, subscription_id, resource_group_name
         location=cluster_location,
         sku=service_bus_sku,
         tags=resource_tag)
-    async_poller = service_bus_client.namespaces.create_or_update(
+    async_poller = service_bus_client.namespaces.begin_create_or_update(
         resource_group_name, service_bus_namespace_name, service_bus_namespace)
     while True:
         async_poller.result(15)