+++ src/scheduled-query/azext_scheduled_query/_params.py
@@ -4,7 +4,7 @@
 # --------------------------------------------------------------------------------------------
 # pylint: disable=line-too-long
 
-from azure.cli.core.commands.parameters import tags_type, get_three_state_flag
+from azure.cli.core.commands.parameters import tags_type, get_three_state_flag, get_enum_type
 from azure.cli.command_modules.monitor.actions import get_period_type
 from azure.cli.command_modules.monitor.validators import get_action_group_validator
 from knack.arguments import CLIArgumentType
@@ -23,6 +23,7 @@ def load_arguments(self, _):
         c.argument('severity', type=int, help='Severity of the alert from 0 (critical) to 4 (verbose).')
         c.argument('window_size', type=get_period_type(), help='Time over which to aggregate metrics in "##h##m##s" format.')
         c.argument('evaluation_frequency', type=get_period_type(), help='Frequency with which to evaluate the rule in "##h##m##s" format.')
+        c.argument('display_name', help='The display name of the alert rule')
         c.argument('condition', options_list=['--condition'], action=ScheduleQueryConditionAction, nargs='+')
         c.argument('condition_query', options_list=['--condition-query'], nargs='+', action=ScheduleQueryConditionQueryAction, help='Query deteils to replace the placeholders in `--condition` argument.')
         c.argument('description', help='Free-text description of the rule.')
@@ -36,3 +37,6 @@ def load_arguments(self, _):
                    options_list=['--mute-actions-duration', '--mad'],
                    help='Mute actions for the chosen period of time (in ISO 8601 duration format) after the alert is fired.')
         c.argument('actions', options_list=['--action', '-a'], action=ScheduleQueryAddAction, nargs='+', validator=get_action_group_validator('actions'))
+        c.argument('auto_mitigate', arg_type=get_three_state_flag(), help='The flag that indicates whether the alert should be automatically resolved or not. The default is true.')
+        c.argument('skip_query_validation', arg_type=get_three_state_flag(), help='The flag which indicates whether the provided query should be validated or not.')
+        c.argument('check_workspace_alerts_storage', options_list=['--check-ws-alerts-storage', '--cwas'], arg_type=get_three_state_flag(), help="The flag which indicates whether this scheduled query rule should be stored in the customer's storage.")