+++ src/aks-preview/HISTORY.md
@@ -2,6 +2,18 @@
 
 Release History
 ===============
+0.5.17
++++++
+* Add parameter '--enable-ultra-ssd' to enable UltraSSD on agent node pool
+
+0.5.16
++++++
+* Vendor SDK using latest swagger with optional query parameter added
+* Support private cluster public fqdn feature
+
+0.5.15
++++++
+* Update to use 2021-05-01 api-version
 
 0.5.14
 +++++